restructured tiles added footer to tiles added removeAdmin checkbox logic if updating user change header to display first name and last name instead of email created profile page started junit testing (TestUserDAO working)

+++ ShopGenius/src/main/java/it/univpm/shopgenius/model/dao/UserDAOImpl.java
@@ -1,16 +1,13 @@
 package it.univpm.shopgenius.model.dao;
 
 import java.util.List;
-import java.util.Set;
 
 import javax.persistence.Query;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
 
-import org.hibernate.Criteria;
 import org.hibernate.Session;
-import org.hibernate.criterion.Restrictions;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Repository;
 
@@ -23,13 +20,15 @@ public class UserDAOImpl extends DefaultDao implements UserDAO {
 
 	private PasswordEncoder passwordEncoder;
 	
-	@Override
-	public User findUserByUsername(String email) {
-		return this.getSession().get(User.class, email);
-	}
+//	@Override
+//	public User findUserByUsername(String email) {
+//		return this.getSession().get(User.class, email);
+//	}
 	
 	@Override
 	public User create(String first_name, String last_name, String email, String password, boolean isEnabled) {
+		if (first_name == null || first_name.length() == 0 || last_name == null || last_name.length() == 0 || email == null || email.length() == 0 || password == null || password.length() == 0)
+			throw new RuntimeException("Empty fields not allowed");
 		User u = new User();
 		u.setFirstName(first_name);
 		u.setLastName(last_name);
@@ -60,7 +59,8 @@ public class UserDAOImpl extends DefaultDao implements UserDAO {
     public void deleteUser(int id) {
     	Session session = this.getSession();
         User user = session.byId(User.class).load(id);
-        session.delete(user);
+        if (user != null)
+        	session.delete(user);
     }
 
     @Override
@@ -104,6 +104,12 @@ public class UserDAOImpl extends DefaultDao implements UserDAO {
 	@Override
 	public void addRole(User user, Role role) {
 		user.addRole(role);
+	}
+
+	@Override
+	public void removeRole(User user, Role role) {
+		user.removeRole(role);
+		
 	}
 
 }
